Initial refactor of decision business logic * Move decision business logic into one dedicated file * No logic change
